feat(importer): Allow use of Any for interfaces
