+++ importers/golang_importer/BUILD.bazel
@@ -5,7 +5,6 @@ go_library(
     srcs = [
         "command.go",
         "go_proto_files_registry.go",
-        "go_to_proto_type.go",
         "golang_importer.go",
     ],
     importpath = "github.com/protoconf/protoconf/importers/golang_importer",
@@ -27,7 +26,6 @@ go_test(
     name = "go_default_test",
     srcs = ["golang_importer_test.go"],
     embed = [":go_default_library"],
-    tags = ["manual"],
     deps = [
         "@com_github_jhump_protoreflect//desc/builder:go_default_library",
         "@com_github_mitchellh_go_homedir//:go_default_library",
